Complete tutorial #7. Add smart enemy.

+++ Broject/src/com/game/main/BasicEnemy.java
@@ -19,7 +19,7 @@ public class BasicEnemy extends GameObject{
 
 	@Override
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 32, 32);
+		return new Rectangle((int) x, (int) y, 32, 32);
 	}
 	
 	@Override
@@ -36,7 +36,7 @@ public class BasicEnemy extends GameObject{
 	@Override
 	public void render(Graphics g) {
 		g.setColor(Color.red);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int) x, (int) y, 16, 16);
 	}
 
 }
