+++ Broject/src/com/game/main/FastEnemy.java
@@ -19,7 +19,7 @@ public class FastEnemy extends GameObject{
 
 	@Override
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 32, 32);
+		return new Rectangle((int) x, (int) y, 32, 32);
 	}
 	
 	@Override
@@ -36,7 +36,7 @@ public class FastEnemy extends GameObject{
 	@Override
 	public void render(Graphics g) {
 		g.setColor(Color.cyan);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int) x, (int) y, 16, 16);
 	}
 
 }
