+++ Broject/src/com/game/main/HUD.java
@@ -5,11 +5,11 @@ import java.awt.Graphics;
 
 public class HUD {
 	
-	public static int HEALTH = 100; 
-	private int greenValue = 255;
+	public static float HEALTH = 100; 
+	private float greenValue = 255;
 	
-	private int score = 0;
-	private int level = 1;
+	private float score = 0;
+	private float level = 1;
 		
 	public void tick() {
 		HEALTH = Game.clamp(HEALTH, 0, 100);
@@ -23,8 +23,8 @@ public class HUD {
 	public void render(Graphics g) {
 		g.setColor(Color.gray);
 		g.fillRect(15, 15, 200, 32);
-		g.setColor(new Color(75, greenValue, 0));
-		g.fillRect(15, 15, HEALTH * 2, 32);		
+		g.setColor(new Color(75, (int) greenValue, 0));
+		g.fillRect(15, 15, (int) HEALTH * 2, 32);		
 		g.setColor(Color.white);
 		g.drawRect(15, 15, 200, 32);
 		
@@ -32,19 +32,19 @@ public class HUD {
 		g.drawString("Level: " + level, 15, 80);
 	}
 	
-	public void score(int score) {
+	public void score(float score) {
 		this.score = score;
 	}
 	
-	public int getScore() {
+	public float getScore() {
 		return score;
 	}
 	
-	public void setLevel(int level) {
+	public void setLevel(float level) {
 		this.level = level;
 	}
 	
-	public int getLevel() {
+	public float getLevel() {
 		return level;
 	}
 }
